requirements.txt
# Web Framework
flask==2.3.3
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
flask-pymongo==2.3.0
werkzeug==2.3.7

# Database
pymongo==4.5.0
pinecone-client==2.2.4

# Document Processing
python-docx==1.0.1
PyPDF2==3.0.1
readability-lxml==0.8.1
beautifulsoup4==4.12.2
docling==0.2.0

# AI and NLP
sentence-transformers==2.2.2
langchain==0.0.325
langchain-openai==0.0.1
langchain-community==0.0.13
openai==1.3.5

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
uuid==1.30

# Web Search
google-api-python-client==2.97.0
